Make separate small button component
refs #49
