+++ components/button-links/types.ts
@@ -1,11 +1,5 @@
-import { ButtonSize } from "../buttons/types";
-
 export interface ButtonLinkProps {
   text: string;
   href: string;
-  /**
-   * Default size is 'base'
-   */
-  size?: ButtonSize;
   dataCy?: string;
 }
