+++ components/buttons/text.tsx
@@ -1,16 +1,15 @@
-import { buttonFontSize, buttonType } from "./functions";
+import { buttonText, buttonType } from "./functions";
 import { ButtonProps } from "./types";
 
 export default function ButtonText(props: ButtonProps) {
-    const fontSize = buttonFontSize(props.size);
+    const text = buttonText(props.loading) ?? props.text;
     const type = buttonType(props.type);
 
     return (
-        <button className={`${fontSize}`}
-            data-cy={props.dataCy ?? "btn-text"}
+        <button data-cy={props.dataCy ?? "btn-text"}
             onClick={props.onClick}
             type={type}>
-            {props.text}
+            {text}
         </button>
     );
 }
