+++ components/buttons/types.ts
@@ -1,6 +1,5 @@
 import { MouseEventHandler } from "react";
 
-export type ButtonSize = "base" | "small";
 export type ButtonType = "button" | "submit" | "reset";
 
 export interface ButtonProps {
@@ -13,10 +12,6 @@ export interface ButtonProps {
    */
   onClick?: MouseEventHandler;
   dataCy?: string;
-  /**
-   * Default size is 'base'
-   */
-  size?: ButtonSize;
   /**
    * Default type is `button`
    */
