+++ stories/buttons/ButtonOutlined.stories.tsx
@@ -10,30 +10,15 @@ const meta: Meta<typeof ButtonOutlined> = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Base: Story = {
+export const Example: Story = {
     args: {
         text: "Base"
     }
 };
 
-export const Small: Story = {
-    args: {
-        text: "Small",
-        size: "small"
-    }
-};
-
-export const BaseLoading: Story = {
+export const Loading: Story = {
     args: {
         text: "Base",
-        loading: true
-    }
-};
-
-export const SmallLoading: Story = {
-    args: {
-        text: "Small",
-        size: "small",
         loading: true
     }
 };
